+++ assets/js/background.js
@@ -1,34 +1,30 @@
-/* global
-
-  chrome
-
-  findQueryParam
-  getOptionsFromStorage
-
-  TRACKERS_BY_ROOT
-  REDIRECT_DATA_BY_TARGET_PARAM
-
-  CHANGE_TYPE_TRACKING_STRIP
-  CHANGE_TYPE_REDIRECT_SKIP
-
-  DEFAULT_STRIPPING_METHOD
-  STRIPPING_METHOD_HISTORY_CHANGE
-  STRIPPING_METHOD_CANCEL_AND_RELOAD
-  STRIPPING_METHOD_BLOCK_AND_RELOAD
-  STRIPPING_METHOD_BLOCK_AND_RELOAD_SKIP_REDIRECTS
-  CONTEXT_MENU_ITEM_ID
-  CONTEXT_MENU_ITEM_TEXT
-
-  STORAGE_KEY_STRIPPING_METHOD_TO_USE
-  STORAGE_KEY_SKIP_KNOWN_REDIRECTS
-  ACTION_OPTIONS_SAVED
-  ACTION_RELOAD_AND_ALLOW_PARAMS
-
-  REASON_UPDATE
-  REASON_INSTALL
-
-*/
 'use strict';
+
+const { findQueryParam, getOptionsFromStorage } = require('./common');
+const { TRACKERS_BY_ROOT }                      = require('./trackers');
+const { REDIRECT_DATA_BY_TARGET_PARAM }         = require('./redirects');
+const {
+  REASON_INSTALL,
+  REASON_UPDATE,
+  STORAGE_KEY_STRIPPING_METHOD_TO_USE,
+  STORAGE_KEY_SKIP_KNOWN_REDIRECTS,
+  ACTION_OPTIONS_SAVED,
+  ACTION_RELOAD_AND_ALLOW_PARAMS,
+  ACTION_GET_STUFF_BY_STRIPPING_METHOD_ID,
+  DEFAULT_STRIPPING_METHOD,
+  STRIPPING_METHOD_HISTORY_CHANGE,
+  STRIPPING_METHOD_CANCEL_AND_RELOAD,
+  STRIPPING_METHOD_BLOCK_AND_RELOAD,
+  STRIPPING_METHOD_BLOCK_AND_RELOAD_SKIP_REDIRECTS,
+  CHANGE_TYPE_REDIRECT_SKIP,
+  CHANGE_TYPE_TRACKING_STRIP,
+  CONTEXT_MENU_ITEM_ID,
+  CONTEXT_MENU_ITEM_TEXT
+}                                               = require('./consts');
+
+
+
+
 
 // What method are we using?  Starts with the default
 let STRIPPING_METHOD_TO_USE = DEFAULT_STRIPPING_METHOD;
@@ -530,6 +526,15 @@ function messageHandler(message, sender, cb) {
       //tabId: DON'T NEED B/C DEFAULT IS CURRENT ACTIVE TAB
       url: message.url
     });
+  }
+
+  // Options page probably wants all the stuff for creating the radio buttons
+  if (message.action === ACTION_GET_STUFF_BY_STRIPPING_METHOD_ID) {
+    // Send it back.
+    cb(STUFF_BY_STRIPPING_METHOD_ID);
+    // Return false so we don't keep the connection open as there's nothing
+    // async happening inside here
+    return false;
   }
 }