+++ assets/js/common.js
@@ -1,21 +1,17 @@
-/* global
-  chrome
-  STORAGE_KEY_STRIPPING_METHOD_TO_USE
+'use strict';
+
+const {
+  STORAGE_KEY_STRIPPING_METHOD_TO_USE,
   DEFAULT_STRIPPING_METHOD
-*/
+}                                       = require('./consts');
 
-// eslint-disable-next-line no-unused-vars
+
 function getOptionsFromStorage(cb, options) {
-  options = options ||
-    {
-      [STORAGE_KEY_STRIPPING_METHOD_TO_USE]:  DEFAULT_STRIPPING_METHOD
-    };
-
+  options = options || { [STORAGE_KEY_STRIPPING_METHOD_TO_USE]: DEFAULT_STRIPPING_METHOD };
   return chrome.storage.sync.get(options, cb);
 }
 
 
-// eslint-disable-next-line no-unused-vars
 function findQueryParam(targetParam, url) {
   url = url || window.location.href;
 
@@ -55,3 +51,9 @@ function findQueryParam(targetParam, url) {
     }
   }
 }
+
+
+module.exports = {
+  getOptionsFromStorage,
+  findQueryParam
+};