+++ assets/js/redirects.js
@@ -13,8 +13,8 @@ const KNOWN_REDIRECTS = [
     patterns: [
       `${SCHEMA}www.google.com/url?`
     ],
-    // Google uses "ping" method sometimes.
-    types: ["main_frame", "ping"]
+    // Google uses 'ping' method sometimes.
+    types: ['main_frame', 'ping']
   },
   {
     // Gmail wraps links in e-mails to pass you through their servers
@@ -23,8 +23,8 @@ const KNOWN_REDIRECTS = [
     patterns: [
       `${SCHEMA}www.google.com/url?`
     ],
-    // I think that for Gmail, "main_frame" is enough.
-    types: ["main_frame"]
+    // I think that for Gmail, 'main_frame' is enough.
+    types: ['main_frame']
   },
   {
     name: 'RedirectingAt',
@@ -32,7 +32,7 @@ const KNOWN_REDIRECTS = [
     patterns: [
       `${SCHEMA}${SUBDOMAIN}.redirectingat.com/?`,
     ],
-    types: ["main_frame"]
+    types: ['main_frame']
   },
   {
     name: 'Facebook',
@@ -41,7 +41,7 @@ const KNOWN_REDIRECTS = [
       `${SCHEMA}l.facebook.com/l.php?`,
       `${SCHEMA}l.messenger.com/l.php?`
     ],
-    types: ["main_frame"]
+    types: ['main_frame']
   },
   {
     name: 'Amazon Affiliate',
@@ -49,7 +49,7 @@ const KNOWN_REDIRECTS = [
     patterns: [
       `${SCHEMA}${SUBDOMAIN}.amazon.ca/gp/redirect.html?`
     ],
-    types: ["main_frame"]
+    types: ['main_frame']
   },
   {
     name: 'Rakuten Marketing',
@@ -57,7 +57,7 @@ const KNOWN_REDIRECTS = [
     patterns: [
       `${SCHEMA}click.linksynergy.com/deeplink?`
     ],
-    types: ["main_frame"]
+    types: ['main_frame']
   },
   {
     name: 'ValueClick',
@@ -66,7 +66,7 @@ const KNOWN_REDIRECTS = [
       `${SCHEMA}www.dpbolvw.net${PATH}?`,
       `${SCHEMA}www.tkqlhce.com${PATH}?`
     ],
-    types: ["main_frame"]
+    types: ['main_frame']
   }
 ];
 
@@ -102,7 +102,7 @@ KNOWN_REDIRECTS.forEach(KNOWN_REDIRECT => {
     };
   }
 
-  // Go through every "type" for this redirect
+  // Go through every 'type' for this redirect
   types.forEach(type => {
     // If we don't already have this type for this target param, add it in
     if (!REDIRECT_DATA_BY_TARGET_PARAM[targetParam].types.includes(type)) {
@@ -167,3 +167,8 @@ function replacePlaceholdersRegex(pattern) {
 function escapeRegExp(str) {
   return str.replace(/[\-\[\]\/\{\}\(\)\*\+\?\.\\\^\$\|]/g, "\\$&");
 }
+
+module.exports = {
+  KNOWN_REDIRECTS,
+  REDIRECT_DATA_BY_TARGET_PARAM
+};